refactor: 주어진 클래스 패키지 이동
