+++ src/main/java/bridge/utils/BridgeMaker.java
@@ -1,6 +1,8 @@
-package bridge;
+package bridge.utils;
 
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
@@ -18,6 +20,9 @@ public class BridgeMaker {
      * @return 입력받은 길이에 해당하는 다리 모양. 위 칸이면 "U", 아래 칸이면 "D"로 표현해야 한다.
      */
     public List<String> makeBridge(int size) {
-        return null;
+        return IntStream.range(0, size)
+                .mapToObj(i -> bridgeNumberGenerator.generate())
+                .map(StringConvertor::convertIntegerToBridgeStatus)
+                .collect(Collectors.toList());
     }
 }
