+++ src/main/java/bridge/utils/BridgeNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.utils;
 
 @FunctionalInterface
 public interface BridgeNumberGenerator {
