+++ src/main/java/bridge/utils/BridgeRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.utils;
 
 import camp.nextstep.edu.missionutils.Randoms;
